requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
aiosqlite==0.21.0
environs==14.2.0
python-jose[cryptography]==3.5.0
pydantic==2.11.7
environs==9.5.0